quick fix

+++ backend/app/models/user.py
@@ -117,7 +117,7 @@ class Profile(db.Model):
     profile_picture_id = db.Column(db.Integer(), db.ForeignKey('media.id'), nullable=True)
     referral_code = db.Column(db.String(255), nullable=True)
     bank = db.Column(db.String(100), nullable=True)
-    account_no = db.Column(db.Integer(12), nullable=True)
+    account_no = db.Column(db.Integer(), nullable=True)
     
     trendit3_user_id = db.Column(db.Integer, db.ForeignKey('trendit3_user.id', ondelete='CASCADE'), nullable=False,)
     trendit3_user = db.relationship('Trendit3User', back_populates="profile")
